+++ R/integrity.R
@@ -98,3 +98,16 @@ validate_repository <- function(x, ..., arg = rlang::caller_arg(x), call = rlang
     cli::cli_abort("{.arg {arg}} must be a valid public repository address. For example, 'FRAMverse/framrsquared' or 'tidyverse/dplyr'. {.arg {x}} is not a valid repository; {.href https://github.com/{x} } does not exist")
   }
 }
+
+github_to_repo_address <- function(x, allow_null = FALSE, arg = rlang::caller_arg(x), call = rlang::caller_env()){
+
+  if(allow_null && is.null(x)){ return(invisible(NULL)) }
+
+  validate_character(x, n = 1, arg = arg, call = call)
+  x = gsub(".*github[.]com[/]", "", x)
+  x = gsub("[/]$", "", x)
+
+  validate_repository(x, arg = arg, call = call)
+
+  return(x)
+}