+++ NEWS.md
@@ -1,5 +1,7 @@
 # codeInspectr 0.0.0.9000 (development version)
 
+* updated all functions that previously required a "repo address" (e.g., "FRAMverse/framrsquared") to also work with the full
+github repository URL (e.g., "https://github.com/FRAMverse/framrsquared/"). Should help avoid confusion, simplify user experience.
 * Added suite of tools for creating PDF versions of package documentation and vignettes. Primary function
 for users is `compile_package_manuals()` (to compile documentation / vignettes for all Github-based R packages stored under a specific user or organization)
 or `build_package_manual()` (to build documentation for a single R package accessible as a Github repository). Additional package dependencies